requirements.txt
alembic==1.11.1
attrs==23.1.0
autopage==0.5.1
brotlipy==0.7.0
certifi==2022.12.7 
cffi==1.15.1 
cfgv==3.3.1
charset-normalizer==2.0.4 
cliff==4.3.0
cmaes==0.9.1
cmd2==2.4.3
cogdl==0.5.3
colorlog==6.7.0
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
distlib==0.3.6
filelock==3.12.0
flake8==6.0.0
flit_core==3.6.0 
fonttools==4.39.4
fsspec==2023.5.0
gensim==3.8.3
grave==0.0.3
greenlet==2.0.2
huggingface-hub==0.14.1
identify==2.5.24
idna==3.4 
importlib-metadata==6.6.0
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
littleutils==0.2.2
llvmlite==0.40.0
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.7.0
mkl-fft==1.3.1
mkl-random 
mkl-service==2.4.0
networkx==3.1
ninja==1.11.1
nodeenv==1.8.0
numba==0.57.0
numpy
ogb==1.3.6
optuna==2.4.0
outdated==0.2.2
packaging==23.1
pandas==2.0.1
pbr==5.11.1
Pillow==9.4.0
platformdirs==3.5.1
pre-commit==3.3.2
prettytable==3.7.0
pycodestyle==2.10.0
pycparser==2.21 
pyflakes==3.0.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.28.1
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
SQLAlchemy==2.0.14
stevedore==5.1.0
tabulate==0.9.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.14
virtualenv==20.23.0
wcwidth==0.2.6
zipp==3.15.0
